requirements.txt
django
psycopg2
djangorestframework
pandas